cloudrun_jobs/enricher/requirements.txt
pyarrow==17.0.0
requests-html==0.10.0
beautifulsoup4==4.12.3
lxml==5.3.0
html5lib==1.1
pyppeteer==2.0.0
nest-asyncio==1.6.0
fake-useragent==1.5.1
pandas==2.2.2
gcsfs==2024.6.0
google-cloud-bigquery==3.25.0
lxml-html-clean==0.1.0
db-dtypes==1.2.0
